feat(backend): refactored report date filter
Ensured that start_date is the start of the filtered day and end_date is
the end of the filtered day.  In theory, the default date filter coming
from the frontend ought to use a complete datetime, but it's only
passing a date.  However, the proxy API requires a datetime.  Prior to
this change it was always setting hours/minutes/seconds to 0.  This
could potentially miss recent reports.