feat(backend): added support for chatbot.report.categories
The category (group) name is not output via the proxy API, only it's id.
Therefore added a ReportCategory model w/ a fixture to store categories
locally.  Then in the API I query those models to work out which
categories map to which reports.  The output is only the group of the
category.  Also provided an API to populate the frontend filter on.
Note that a report can have multiple categories.