feat(backend): added local scheulder service to run cron jobs
Added local scheduler service to run cron jobs.  This mimics the
heroku-scheduler dyno that is run in prduction.  The scheduled task
being run is the purge_camera_media docker management command.  However,
that has to be wrapped in a calling script which sets certain
environment variables since cron runs in such a diminished environment.