feat(backend): added info attribute to MapReqeust response
Added a new field to store group_info on a DataType.  And a new property
on MapRequest to output the group and group_info of that request
(remember that all DataTypes of a MapRequest are guaranteed to have the
same group and group_info).  Then updated the groupby function in the
MapRequestListSerializer.to_representation to use the new field.  Also
updated the DataTypes fixture to include group_info.