feat(backend): added backend support for long timeseries
The GetTimeSeries request of geoserver can only cope w/ 100 timestamps
and some of the subseasonal have many more.  This PR "chunks" them into
subsets of 100 and passes an array of URLs for the frontend to query.
The corresponding frontend work to merge the CSV returned by those
queries together still needs to be done.