feat(backend): fixed invalid JSON error when parsing camera media messages w/ location_info
When determining the geometry for a camera media object I added a geos
object into the message body to use later.  Unfortunately, that object
is invalid JSON when it comes time to parse it again in case I have to
create an alert based on that media object.

To fix this I create the geos object separately from the message body.