feat(backend): remove special case for DataType 35006
DataType 35006 is no longer needed for "Wildfire Simulation"
MapRequests.  It was always associated w/ _creating_ a request and never
w/ any of the resultant request layers.  Therefore there was some code
to treat it specially.  This PR removes that code since it will no
longer form part of the MapRequest - it will be removed from both the
backend database and the frontend form.