chore(backend): updated dj-rest-auth settings
When I upgraded dependencies I found that dj-rest-auth had made some
breaking changes.  I wound up having to update the way that
dj-rest-auth's settings are configured.  And explicitly define the
`create_token` function.  In the long-term most of this code will be
removed from safers as part of a general refactoring of authentication.
But for now I couldn't run the current code without making these
changes.