feat(backend): made the alert response a FeatureCollection
This code ensures Alerts are serialized as FeatureCollections rather
than the frontend having to transofrm them.  The center property is used
as the geometry of each feature.  And the geometry_collection is
included in the properties so we still have access to the orginial
geometry (ie: an alert could span multiple shapes w/in a
geometry_collection).