feat(backend): prevent duplicate favorite items
The way I was computing favorite items was error-prone.  I reverted to
the less fancy but recommended way:

First compute a user's favorite items and save that in a list.
Then add an annotation to all items on whether or not an item is in that
list.
Then order by that annotation.