feat(backend): prevent updating all user fields
When editing a UserProfile in the frontend users can currently change
their organization and role.  Passing this onto the proxy API (which in
turn calls FusionAuth) causes errors.  And, anyway, it doesn't make
sense to allow a user to change their organization or to change their
(own) role.

This PR adds a new serializer to use when updating a user.  The fields
are the same (so that the content passed to the proxy API has the same
structure).  But updating the organization and role is disabled; those
are read-only fields.

To avoid confusion, the ability to modify these fields in the frontend
UserProfile form should also be disabled.