feat(backend): remove UserProfile model (#322)
* feat(backend): remove UserProfile model

Removing UserProfile model.  This PR deletes the admins, serializers, and
views, signals, etc.  associated w/ profiles.  It also removes the 1to1
relationship between profiles and users in one migration, then completely 
removes the model in another migration.  Finally, a third migration replaces
the model w/ a JSONField.  That field can be synchronised to/from the gateway.  
I still need to write the actual synchronisation function, but that must wait until
after the authentication mechanism is refactored and there is a gateway client to use.