feat(backend): added CachedTransientModelManager
Added a ModelManager and QuerySet that allows models to be retrieved
from remote APIs but still treats them internally as a django model w/
serializers, views, etc.

Also added a class to act as a safers gateway client.  Initially this
will be used for the remote API when retrieving organizations.  But in
the long term I will replace all explicit calls to the requests library
with this class.