fix(backend): Ensure cache tables created
When starting from scratch, the backend was trying to access tables
before they were created e.g. during the migration phase, but at
that point, the cache tables didn't yet exist. I switched the cache
tables creation to be the first thing in the script and it all now
seems to work as expected.