feat(backend): change the way alert images are referenced
The original code stored the remote_url of any camera images associated
with an alert in the database.  Unfortunately, those URLs could
potentially expire or else be deleted.  A subsequent PR ensured that the
local_url was used as well as a local_url to a thumbnail image.
However, the storage backend (AWS S3 - served by bucketeer) only
provides presigned URLs which can also potentially expire.  This PR no
longer stores any hard-coded URL in the database for an alert image.
Instead it uses a model property to compute the URLs for images and
thumbnails on-demand with each view.